% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{cols}
\alias{cols}
\title{Break html slide into columns}
\usage{
cols(side = "l", w = 50)
}
\arguments{
\item{side}{"l" or "r"}

\item{w}{width of slide, in percent}
}
\description{
Use this in a chunk to create handy html code, but remember to have results = "asis"
}
