% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{pars}
\alias{pars}
\title{Establish Elie-style graphic pars}
\usage{
pars(...)
}
\arguments{
\item{...}{additional arguments for par.}
}
\description{
I like these:
- mar = c(3,3,2,2),
- bty = "l",
- mgp = c(1.5,.25,0),
- tck = 0.01,
- cex.lab = 1.2
}
