% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{pics}
\alias{pics}
\title{Insert image}
\usage{
pics(file, w = 100)
}
\arguments{
\item{file}{relative file location}

\item{w}{width of image, in percentage}
}
\description{
Use this in a chunk to create handy html code, but remember to have results = "asis"
}
